+++ kfinance/integrations/tool_calling/all_tools.py
@@ -35,9 +35,8 @@ from kfinance.domains.line_items.line_item_tools_va import (
     GetVisibleAlphaFinancialLineItemFromIdentifiers,
 )
 from kfinance.domains.mergers_and_acquisitions.merger_and_acquisition_tools import (
-    GetAdvisorsForCompanyInTransactionFromIdentifier,
-    GetMergerInfoFromTransactionId,
     GetMergersFromIdentifiers,
+    GetMergersInfoFromTransactionIds,
 )
 from kfinance.domains.prices.price_tools import (
     GetHistoryMetadataFromIdentifiers,
@@ -104,8 +103,7 @@ ALL_TOOLS: list[type[KfinanceTool]] = [
     # Statements
     GetFinancialStatementFromIdentifiers,
     # Mergers & Acquisitions
-    GetAdvisorsForCompanyInTransactionFromIdentifier,
-    GetMergerInfoFromTransactionId,
+    GetMergersInfoFromTransactionIds,
     GetMergersFromIdentifiers,
     # Rounds of Funding
     GetRoundsOfFundingFromIdentifiers,